% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runa.R
\name{runa}
\alias{runa}
\title{Test function to work with Bioconductor rhdf5}
\usage{
runa(...)
}
\arguments{
\item{...}{ignored}
}
\value{
message about Bioconductor
}
\description{
Just a message to test the rhdf5 import.
}
\examples{
runa()
}
